+++ TwitchLib.Client.Models/WhisperMessage.cs
@@ -1,6 +1,5 @@
 using System.Drawing;
 using TwitchLib.Client.Enums;
-using TwitchLib.Client.Models.Common;
 using TwitchLib.Client.Models.Internal;
 
 namespace TwitchLib.Client.Models
